+++ .adonisjs/client/registry/schema.d.ts
@@ -175,6 +175,30 @@ export interface Registry {
       errorResponse: ExtractErrorResponse<Awaited<ReturnType<import('#controllers/budget_controller').default['show']>>> | { status: 422; response: { errors: SimpleError[] } }
     }
   }
+  'boats.portStays.store': {
+    methods: ["POST"]
+    pattern: '/boats/:id/port-stays'
+    types: {
+      body: ExtractBody<InferInput<(typeof import('#validators/boat_port_stay_validator').boatPortStayValidator)>>
+      paramsTuple: [ParamValue]
+      params: { id: ParamValue }
+      query: ExtractQuery<InferInput<(typeof import('#validators/boat_port_stay_validator').boatPortStayValidator)>>
+      response: ExtractResponse<Awaited<ReturnType<import('#controllers/boat_port_stay_controller').default['store']>>>
+      errorResponse: ExtractErrorResponse<Awaited<ReturnType<import('#controllers/boat_port_stay_controller').default['store']>>> | { status: 422; response: { errors: SimpleError[] } }
+    }
+  }
+  'boats.portStays.destroy': {
+    methods: ["DELETE"]
+    pattern: '/boats/:id/port-stays/:stayId'
+    types: {
+      body: {}
+      paramsTuple: [ParamValue, ParamValue]
+      params: { id: ParamValue; stayId: ParamValue }
+      query: {}
+      response: ExtractResponse<Awaited<ReturnType<import('#controllers/boat_port_stay_controller').default['destroy']>>>
+      errorResponse: ExtractErrorResponse<Awaited<ReturnType<import('#controllers/boat_port_stay_controller').default['destroy']>>>
+    }
+  }
   'boats.edit': {
     methods: ["GET","HEAD"]
     pattern: '/boats/:id/edit'